backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
mysql-connector-python==8.2.0
pandas==2.1.3
openpyxl==3.1.2
python-dotenv==1.0.0
bcrypt==4.1.1
Werkzeug==3.0.1

